requirements.txt
streamlit
google-api-python-client
python-dotenv
pandas
numpy
requests
beautifulsoup4
lxml
gnews
git+https://github.com/FinanceData/FinanceDataReader.git